+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2.3.0 - 2025-11-11
+### Added
+- Added an option to override image alt text and description.
+- Added the functionality of sorting images in the field.
+- Fixed "Sync metadata" utility.
+- Added the option to use imageshop images foropengraph, with SEOmatic plugin.
+
 ## 2.1.0 - 2024-12-19
 ### Added
 - Sync utility to update existing fields.